feat: Implement GhostMode and Retry Button
Added GhostMode funcionality with a UI ResourceBar. I made two scripts, one for the GhostMode itself and another one to manage the resource bar UI.

Also finally added the funcionality of reloading the scene to the retry button.

Fixed minor problems to animations, level end and bullets, which now dissapear with distance.